web-anaylzer/requirements/base.txt
# Core dependencies
PyQt5==5.15.7
mysql-connector-python==8.0.28
pyyaml==6.0
python-dateutil==2.8.2
watchdog==2.1.6
numpy==1.23.5
pandas==1.5.2
matplotlib==3.6.2
pyqtgraph==0.13.3
scipy==1.10.1
statsmodels==0.14.0
user-agents==2.2.0
Jinja2==3.1.2